split console output to terminal
