+++ tests/basic/test_editblock.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock, patch
 from aider.coders import Coder
 from aider.coders import editblock_coder as eb
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.terminal import Terminal
 from aider.models import Model
 
 
@@ -297,7 +297,7 @@ These changes replace the `subprocess.run` patches with `subprocess.check_output
         files = [file1]
 
         # Initialize the Coder object with the mocked IO and mocked repo
-        coder = Coder.create(self.GPT35, "diff", io=InputOutput(), fnames=files, pretty=False)
+        coder = Coder.create(self.GPT35, "diff", io=Terminal(), fnames=files, pretty=False)
 
         def mock_send(*args, **kwargs):
             coder.partial_response_content = f"""
@@ -337,7 +337,7 @@ new
         coder = Coder.create(
             self.GPT35,
             "diff",
-            io=InputOutput(dry_run=True),
+            io=Terminal(dry_run=True),
             fnames=files,
             dry_run=True,
             pretty=False,
